+++ charger/easee/types.go
@@ -151,9 +151,9 @@ type CircuitSettings struct {
 	DynamicCircuitCurrentP1    *float64 `json:"dynamicCircuitCurrentP1,omitempty"`
 	DynamicCircuitCurrentP2    *float64 `json:"dynamicCircuitCurrentP2,omitempty"`
 	DynamicCircuitCurrentP3    *float64 `json:"dynamicCircuitCurrentP3,omitempty"`
-	MaxCircuitCurrentP1        *int     `json:"maxCircuitCurrentP1,omitempty"`
-	MaxCircuitCurrentP2        *int     `json:"maxCircuitCurrentP2,omitempty"`
-	MaxCircuitCurrentP3        *int     `json:"maxCircuitCurrentP3,omitempty"`
+	MaxCircuitCurrentP1        *float64 `json:"maxCircuitCurrentP1,omitempty"`
+	MaxCircuitCurrentP2        *float64 `json:"maxCircuitCurrentP2,omitempty"`
+	MaxCircuitCurrentP3        *float64 `json:"maxCircuitCurrentP3,omitempty"`
 	EnableIdleCurrent          *bool    `json:"enableIdleCurrent,omitempty"`
 	OfflineMaxCircuitCurrentP1 *int     `json:"offlineMaxCircuitCurrentP1,omitempty"`
 	OfflineMaxCircuitCurrentP2 *int     `json:"offlineMaxCircuitCurrentP2,omitempty"`
